# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Haushaltsbuch
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /action/{id}:
    # our controller name
    x-swagger-router-controller: action
    get:
      operationId: getOne
      description: get an action
      # define the type of response for Success "200" and Error
      parameters:
        - name: id
          type: number
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Action"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      operationId: update
      description: update an action
      parameters:
        - name: id
          description: Action id
          type: number
          in: path
          required: true
        - name: action
          description: Action properties
          in: body
          required: true
          schema:
            $ref: "#/definitions/ActionNoId"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      operationId: deleteById
      description: delete an action
      parameters:
        - name: id
          description: Action id
          type: number
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  UserSecure:
    type: object
    properties:
      id: 
        type: number
      name:
        type: string
  Action:
    type: object
    properties:
      id:
        type: number
      value:
        type: number
        format: double
      date:
        type: string
        format: date-time
      interval:
        type: integer
      category:
        $ref: "#/definitions/Category"
      userPayed: 
        $ref: "#/definitions/UserSecure"
      userBought: 
        $ref: "#/definitions/UserSecure"
  ActionNoId:
    type: object
    properties:
      value:
        type: number
        format: double
      date:
        type: string
        format: date-time
      interval:
        type: integer
      category:
        $ref: "#/definitions/Category"
      userPayed: 
        $ref: "#/definitions/UserSecure"
      userBought: 
        $ref: "#/definitions/UserSecure"
...
  Category:
    type: object
    properties:
      id:
        type: number
      value:
        type: string
  GeneralResponse:
    type: object
    properties:
      success:
        type: number
        description: returns 1 if successful
      description:
        type: string
        description: a short comment 
    required:
      - success
      - description
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string